requirements.txt
numpy~=1.18.3
matplotlib~=3.3.2
Pillow~=8.0.1
scipy~=1.4.1